python-embeddings/requirements.txt
fastapi
uvicorn
numpy==1.26.4
faiss-cpu
sentence-transformers==2.7.0
torch==2.2.2
langchain
langchain_text_splitters